+++ index.html
@@ -405,7 +405,7 @@
       </div>
     </section>
   </template>
-<script src="js/main.js"></script>
+
 </body>
     </footer>
 
@@ -459,6 +459,6 @@
         </div>
       </section>
     </template>
-    <script src="js/functions.js"></script>
+    <script src="./js/main.js" type="module"></script>
   </body>
 </html>
